Move plugin bundle projection to a follow-up PR
