+++ src/coder_eval/models/__init__.py
@@ -23,14 +23,12 @@ from coder_eval.models.container_paths import (
     AGENT_HOME,
     AGENT_UID,
     AGENT_USERNAME,
-    CONTAINER_AGENT_SKILLS_DIR,
     CONTAINER_AGENT_WORK_DIR,
     CONTAINER_CLAUDE_SHIM,
     CONTAINER_DROP_SHIM,
     CONTAINER_GRADER_DIR,
     CONTAINER_INPUT_DIR,
     CONTAINER_OUTPUT_DIR,
-    CONTAINER_PRIVATE_PLUGIN_DIR,
     CONTAINER_TASK_DIR,
     CONTAINER_WORK_DIR,
     RESERVED_CONTAINER_DIRS,
@@ -279,14 +277,12 @@ __all__ = [  # noqa: RUF022 - Keep grouped by category for readability
     "AGENT_HOME",
     "AGENT_UID",
     "AGENT_USERNAME",
-    "CONTAINER_AGENT_SKILLS_DIR",
     "CONTAINER_AGENT_WORK_DIR",
     "CONTAINER_CLAUDE_SHIM",
     "CONTAINER_DROP_SHIM",
     "CONTAINER_GRADER_DIR",
     "CONTAINER_INPUT_DIR",
     "CONTAINER_OUTPUT_DIR",
-    "CONTAINER_PRIVATE_PLUGIN_DIR",
     "CONTAINER_TASK_DIR",
     "CONTAINER_WORK_DIR",
     "RESERVED_CONTAINER_DIRS",
